Moved `Has` traits to own module
Closes https://github.com/PistonDevelopers/graphics/issues/659
